backend_venue/requirements.txt
asgiref==3.8.1
Django==5.1.4
sqlparse==0.5.3
tzdata==2024.2
psycopg2-binary==2.9.9
dj-rest-auth==4.0.0
django-allauth==0.52.0
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
pillow==10.2.0
channels==4.0.0
daphne==4.0.0
Pillow
python-decouple
dj-rest-auth
gunicorn==21.2.0
python-dotenv==1.0.1